+++ CdbBack/Cdb.Domain/DTO/CdbYeldDTO.cs
@@ -2,7 +2,7 @@
 {
     public class CdbYeldDTO
     {
-        public double GrossAmount { get; set; }
-        public double NetAmount { get; set; }
+        public decimal GrossAmount { get; set; }
+        public decimal NetAmount { get; set; }
     }
 }
